requirements.txt
streamlit
pdfplumber
requests
python-dotenv
sentence-transformers
torch